feat: remove redundant files
