+++ frontend/src/pages/Guard/components/CurrentStatus.tsx
@@ -1,32 +1,6 @@
-// pages/Guard/index.tsx
-
-// import { styled } from '@mui/material/styles';
 import ParkingLot from '../../../components/ParkingLot';
 import SelectedLotMenu from '../../../components/SelectedLotMenu';
 import SelectedFloorMenu from '../../../components/SelectedFloorMenu';
-
-
-// import Button from '@mui/material/Button';
-
-
-// const CommonButton = styled(Button)`
-//   text-transform: none;
-//   font-size: 16px;
-//   padding: 10px 15px;
-//   border: 2px solid;
-//   border-radius: 12px;
-//   background-color: #ffffff;
-//   border-color: #3B88C3;
-//   &:hover {
-//     border-color: blue-dark;
-//     background-color: #ffffff;
-//     box-shadow: 0 0 0 0.1rem #3B88C3;
-//   }
-//   &:focus {
-//     box-shadow: 0 0 0 0.2rem #3B88C3;
-//   }
-//   font-family: 'Lexend', sans-serif;
-// `;
 
 
 export default function CurrentStatus() {
@@ -40,7 +14,7 @@ export default function CurrentStatus() {
           <SelectedFloorMenu />
         </div>
       </div>
-      {/* <Divider className='text-blue-dark' sx={{ width: '80%', borderColor: '#3B88C3', borderWidth: '1.5px', borderRadius: '24px' }} /> */}
+
       <div className='flex w-full max-h-full mt-2  justify-center'>
         < ParkingLot floor={1} slot={1} modal={true} carlicense={null} usagehistory={false} />
       </div>
