+++ frontend/src/pages/Guard/components/UsageHistory.tsx
@@ -2,8 +2,6 @@
 
 import SelectedLotMenu from '../../../components/SelectedLotMenu';
 import SelectedFloorMenu from '../../../components/SelectedFloorMenu';
-
-// import Divider from '@mui/material/Divider';
 // import { useNavigate } from "react-router-dom";
 
 
@@ -12,11 +10,6 @@ import ParkingLot from '../../../components/ParkingLot';
 // const commonButtonClass = "text-white bg-blue-dark hover:bg-blue-exdark focus:outline-none focus:ring-2 focus:ring-blue-exdark rounded-full text-md px-5 py-2.5 text-center me-2 mt-6 flex";
 
 export default function UsageHistory() {
-
-  // const navigate = useNavigate();
-  // const handleClick = () => {
-  //   navigate("/guard/usagehistory/[spaceID]");
-  // };
 
   return (
     <div className='flex flex-col items-center justify-start'>
@@ -27,7 +20,6 @@ export default function UsageHistory() {
           <SelectedFloorMenu />
         </div>
       </div>
-      {/* <Divider className='text-blue-dark' sx={{ width: '80%', borderColor: '#3B88C3', borderWidth: '1.5px', borderRadius: '24px' }} /> */}
       <div className='flex w-full sm:max-h-96  justify-center'>
         < ParkingLot floor={1} slot={1} modal={false} carlicense={null} usagehistory={true} />
       </div>
