+++ frontend/src/pages/Guard/components/modals/UsageDurationModal.tsx
@@ -1,5 +1,3 @@
-// import { useState } from "react";
-
 import Backdrop from '@mui/material/Backdrop';
 import Box from '@mui/material/Box';
 import Modal from '@mui/material/Modal';
