--- frontend/src/services/addUserService.ts
@@ -1,6 +1,0 @@
-export class AddUserService {
-  public addUser(id: string, password: string, role: string): boolean {
-    // PATCH /staff/users/:id
-    return true;
-  }
-}
