+++ frontend/src/services/userService.ts
@@ -12,8 +12,27 @@ export class UserService {
       status: "common",
     };
   }
+  
+  // Change Personal Info
   public updateUserInfo(newInfo: userInfo): boolean {
-    // TODO: update user info
+    // TODO: update user info    
+    // PATCH /staff/users/:id
     return newInfo.id !== "";
+  }
+
+  // Add user 
+  public addUser(id: string, password: string, role: string): boolean {
+    // POST /users
+    return true;
+  }
+
+  // Register
+  public registerCarOwner(id: string, name: string, carId: string, phone: string, email: string, password: string): boolean {
+    // PATCH /staff/users/:id
+    return true;
+  }
+  public registerGuard(id: string, name: string, password: string): boolean {
+    // PATCH /staff/users/:id
+    return true;
   }
 }
