+++ 進度.txt
@@ -1,0 +1,32 @@
+11/17 
+移除 postcss.config.js
+components: 增加 loading
+contexts: 增加 toast
+public: 增加logo.svg(方的、原始)、logo.png、favicon.ico、
+
+11/24
+三個route  開始、身分、登出
+TODO: ratio、duration、revservation
+
+11/24
+components 新增三個modal、一個guard headerbar
+ratio選哪個icon?
+
+11/25
+完成usage duration
+切換到哪一頁，主題名字秀在icon旁邊(手機版本)?
+
+11/26
+Todo: 整合、settings page、navbar(加logout)、(api)、history 加上Modal
+
+11/27
+鄭: Done
+
+11/30
+
+
+待解決: 
+動畫速度調整??
+Usage Detail(是自己的車牌號碼的話 車子顏色變黃色)
+current status modal內容
+下拉式選單寬度不對
